Update README.md
Hello! I'm the guy who wrote the original url2pdf package. Anyway, thanks for your work modernising it.!

I just wanted to update the readme for the renderHTML example. The current example says it will return a path, but in reality it returns a buffer. I tried to keep suggested change in the same style as the previous example (vanilla es5). I have also tested it and it saves the pdf correctly.